Rewrite print_time into TimePrinter
