--- src/modules/print_time.h
@@ -1,15 +1,0 @@
-#ifndef  __swaystatus_print_time_H__
-# define __swaystatus_print_time_H__
-
-# ifdef __cplusplus
-extern "C" {
-# endif
-
-void init_time(void *config);
-void print_time();
-
-# ifdef __cplusplus
-}
-# endif
-
-#endif
